feat: prevent duplication when join & leave room #27

+++ src/rooms/users/users.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaClient } from '@prisma/client';
+import { PrismaClient, Room, User } from '@prisma/client';
 import { RoomsService } from '../rooms.service';
 import { VotesService } from 'src/votes/votes.service';
 import { UsersService } from 'src/users/users.service';
@@ -16,45 +16,53 @@ export class RoomUsersService {
   async create(roomId: string, userFirebaseId: string) {
     const user = await this.usersService.findOne(userFirebaseId, 'firebase');
 
-    const room = await this.prisma.room.update({
-      where: { id: roomId },
-      data: {
-        users: {
-          connect: { firebaseId: userFirebaseId },
+    if (!this.hasAlreadyEntered(roomId, user)) {
+      const room = await this.prisma.room.update({
+        where: { id: roomId },
+        data: {
+          users: {
+            connect: { firebaseId: userFirebaseId },
+          },
+          participantsLength: {
+            increment: 1,
+          },
         },
-        participantsLength: {
-          increment: 1,
-        },
-      },
-      include: { users: true, plapo: true },
-    });
+        include: { users: true, plapo: true },
+      });
 
-    await this.votesService.create({
-      userId: user.id,
-      plapoId: room.plapo.id,
-    });
+      await this.votesService.create({
+        userId: user.id,
+        plapoId: room.plapo.id,
+      });
+    }
 
-    return await this.roomsService.findOne(room.id);
+    return await this.roomsService.findOne(roomId);
   }
 
   async delete(roomId: string, userFirebaseId: string) {
     const user = await this.usersService.findOne(userFirebaseId, 'firebase');
 
-    const room = await this.prisma.room.update({
-      where: { id: roomId },
-      data: {
-        users: {
-          disconnect: { firebaseId: userFirebaseId },
+    if (this.hasAlreadyEntered(roomId, user)) {
+      const room = await this.prisma.room.update({
+        where: { id: roomId },
+        data: {
+          users: {
+            disconnect: { firebaseId: userFirebaseId },
+          },
+          participantsLength: {
+            decrement: 1,
+          },
         },
-        participantsLength: {
-          decrement: 1,
-        },
-      },
-      include: { users: true },
-    });
+        include: { users: true, plapo: true },
+      });
 
-    await this.votesService.delete(user.votes[0].id);
+      await this.votesService.delete(user.votes[0].id);
+    }
 
-    return await this.roomsService.findOne(room.id);
+    return await this.roomsService.findOne(roomId);
+  }
+
+  hasAlreadyEntered(roomId: string, user: User & { rooms: Room[] }): boolean {
+    return user.rooms.filter((room) => room.id == roomId).length != 0;
   }
 }
